refactor : member의 service, controller 수정.
