+++ src/main/java/com/study/yesmarket/member/exception/NotMatchedIdException.java
@@ -1,11 +1,10 @@
 package com.study.yesmarket.member.exception;
 
 import com.study.yesmarket.common.exception.BusinessException;
-import com.study.yesmarket.login.exception.LoginErrorCode;
 
 public class NotMatchedIdException extends BusinessException {
 
     public NotMatchedIdException() {
-        super(LoginErrorCode.NotMatchedId);
+        super(MemberErrorCode.NotMatchedId);
     }
 }
